+++ pkg/backend/display/display.go
@@ -42,8 +42,6 @@ func ShowEvents(
 	}
 
 	if opts.JSONDisplay {
-		// TODO[pulumi/pulumi#2390]: enable JSON display for real deployments.
-		contract.Assertf(isPreview, "JSON display only available in preview mode")
 		ShowJSONEvents(op, action, events, done, opts)
 		return
 	}